requirements.txt
aiodns==3.5.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-retry==2.9.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
async-timeout==5.0.1
attrs==25.3.0
audioread==3.0.1
backoff==2.2.1
bcrypt==4.3.0
blis==1.3.0
boto3==1.40.42
botocore==1.40.42
Brotli==1.1.0
catalogue==2.0.10
certifi==2025.8.3
cffi==2.0.0
cfgv==3.4.0
charset-normalizer==3.4.3
chatterbox-tts==0.1.3
click==8.1.8
cloudpathlib==0.22.0
colorama==0.4.6
confection==0.1.5
conformer==0.3.2
cryptography==45.0.7
cymem==2.0.11
Cython==3.1.4
decorator==5.2.1
Deprecated==1.2.18
diffusers==0.29.0
distlib==0.4.0
dnspython==2.7.0
einops==0.8.1
email-validator==2.3.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl
exceptiongroup==1.3.0
fastapi==0.116.2
fastapi-cli==0.0.12
fastapi-cloud-cli==0.1.5
filelock==3.19.1
frozenlist==1.7.0
fsspec==2025.9.0
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.35.0
identify==2.6.14
idna==3.10
importlib_metadata==8.7.0
inquirerpy==0.3.4
invoke==2.2.0
itsdangerous==2.2.0
jaconv==0.4.0
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.2
langcodes==3.5.0
language_data==1.3.0
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.43.0
marisa-trie==1.3.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml_dtypes==0.5.3
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.4
murmurhash==1.0.13
networkx==3.2.1
nodeenv==1.9.1
numba==0.60.0
numpy==2.0.2
onnx==1.19.0
orjson==3.11.3
packaging==25.0
paramiko==4.0.0
pfzy==0.3.4
pillow==11.3.0
pkuseg==0.0.25
platformdirs==4.4.0
pooch==1.8.2
pre_commit==4.3.0
preshed==3.0.10
prettytable==3.16.0
prompt_toolkit==3.0.52
propcache==0.3.2
protobuf==6.32.1
py-cpuinfo==9.0.0
pycares==4.11.0
pycparser==2.23
pydantic==2.11.9
pydantic-extra-types==2.10.5
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
pykakasi==2.3.0
PyNaCl==1.6.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
python-telegram-bot==22.5
PyYAML==6.0.2
regex==2025.9.1
requests==2.32.5
resemble-perth==1.0.1
rich==14.1.0
rich-toolkit==0.15.1
rignore==0.6.4
runpod==1.7.13
s3tokenizer==0.2.0
s3transfer==0.14.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.13.1
sentry-sdk==2.38.0
shellingham==1.5.4
six==1.17.0
smart_open==7.3.1
sniffio==1.3.1
soundfile==0.13.1
soxr==1.0.0
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
starlette==0.48.0
sympy==1.13.1
thinc==8.3.6
threadpoolctl==3.6.0
tokenizers==0.20.3
tomli==2.2.1
tomlkit==0.13.3
torch==2.6.0
torchaudio==2.6.0
tqdm==4.67.1
tqdm-loggable==0.2
transformers==4.46.3
typer==0.17.4
typing-inspection==0.4.1
typing_extensions==4.15.0
ujson==5.11.0
urllib3==1.26.20
uvicorn==0.35.0
uvloop==0.21.0
virtualenv==20.34.0
wasabi==1.1.3
watchdog==6.0.0
watchfiles==1.1.0
wcwidth==0.2.13
weasel==0.4.1
websockets==15.0.1
wrapt==1.17.3
yarl==1.20.1
zipp==3.23.0
